refactor(component): move Legend and Description components
