+++ packages/big-design/src/components/Fieldset/Description/Description.tsx
@@ -1,0 +1,5 @@
+import React, { memo } from 'react';
+
+import { Small, TextProps } from '../../Typography';
+
+export const FieldsetDescription: React.FC<TextProps> = memo(({ className, style, ...props }) => <Small {...props} />);
