+++ packages/big-design/src/components/Fieldset/Description/index.ts
@@ -1,0 +1,1 @@
+export { FieldsetDescription } from './Description';
