+++ packages/big-design/src/components/Fieldset/Legend/Legend.tsx
@@ -1,0 +1,9 @@
+import React, { memo } from 'react';
+
+import { HeadingProps } from '../../Typography';
+
+import { StyledFieldsetLegend } from './styled';
+
+export const FieldsetLegend: React.FC<HeadingProps> = memo(({ className, style, ...props }) => (
+  <StyledFieldsetLegend {...props} />
+));
