+++ packages/big-design/src/components/Fieldset/Legend/index.ts
@@ -1,0 +1,1 @@
+export { FieldsetLegend } from './Legend';
