+++ packages/big-design/src/components/Fieldset/Legend/styled.tsx
@@ -1,0 +1,12 @@
+import { theme as defaultTheme } from '@bigcommerce/big-design-theme';
+import styled, { DefaultTheme, StyledComponent } from 'styled-components';
+
+import { StyleableH3 } from '../../Typography/private';
+
+export const StyledFieldsetLegend = styled(StyleableH3).attrs({ as: 'legend' })`
+  &:not(:last-child) {
+    margin-bottom: ${({ theme }) => theme.spacing.xxSmall};
+  }
+` as StyledComponent<'legend', DefaultTheme>;
+
+StyledFieldsetLegend.defaultProps = { theme: defaultTheme };
