+++ packages/big-design/src/components/Fieldset/styled.tsx
@@ -1,7 +1,5 @@
 import { theme as defaultTheme } from '@bigcommerce/big-design-theme';
-import styled, { DefaultTheme, StyledComponent } from 'styled-components';
-
-import { StyleableH3, StyleableSmall } from '../Typography/private';
+import styled from 'styled-components';
 
 export const StyledFieldset = styled.fieldset`
   border: none;
@@ -13,14 +11,4 @@ export const StyledFieldset = styled.fieldset`
   }
 `;
 
-export const StyledFieldsetDescription = styled(StyleableSmall)``;
-
-export const StyledFieldsetLegend = styled(StyleableH3).attrs({ as: 'legend' })`
-  &:not(:last-child) {
-    margin-bottom: ${({ theme }) => theme.spacing.xxSmall};
-  }
-` as StyledComponent<'legend', DefaultTheme>;
-
 StyledFieldset.defaultProps = { theme: defaultTheme };
-StyledFieldsetLegend.defaultProps = { theme: defaultTheme };
-StyledFieldsetDescription.defaultProps = { theme: defaultTheme };
